Extract ephemNavConverter from message types
